feat: Update icons in fnpm-ui to use Tabler Icons
This commit updates the icons used in the fnpm-ui package to use Tabler Icons instead of Lucide Icons. The Lucide Icons package has been removed as a dependency in the package.json file.

The switch to Tabler Icons provides a wider range of icons and better compatibility with the current version of React.

Closes #123